+++ src/Cabinet/Common/GetConfig.php
@@ -7,9 +7,9 @@ use Haikiri\MikBiLL\ResponseWrapper;
 class GetConfig extends ResponseWrapper
 {
 
-	public function getCompanyInfo(): GetCompanyInfoModel
+	public function getCompanyInfo(): GetCompanyInfo
 	{
-		return new GetCompanyInfoModel($this->getData("company_info"));
+		return new GetCompanyInfo($this->getData("company_info"));
 	}
 
 }
